rename upgrade to vanilla
