+++ app/upgrades/vanilla/upgrades.go
@@ -1,4 +1,4 @@
-package v1
+package vanilla
 
 import (
 	"context"
